+++ .editorconfig
@@ -1,0 +1,20 @@
+# EditorConfig is awesome: http://EditorConfig.org
+
+# top-most EditorConfig file
+root = true
+
+# Unix-style newlines with a newline ending every file
+[*]
+end_of_line = lf
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*.{php,html5,twig,yml,json}]
+indent_style = space
+indent_size = 4
+
+[*.{svg,min.css,min.js}]
+insert_final_newline = false
+
+[*.md]
+trim_trailing_whitespace = false
