+++ public/autosuggester.1896a185.css
@@ -1,0 +1,2 @@
+.autosuggester-input-mirror{left:-9999px!important;overflow:auto!important;position:absolute!important;top:0!important;white-space:pre-wrap!important;word-wrap:break-word!important}.autosuggester-box-container{position:absolute;z-index:9}.autosuggester-box{background-color:var(--content-bg,#fff);border:1px solid var(--content-border,#bbb);border-radius:var(--border-radius,0 0 3px 3px);box-shadow:0 3px 6px 0 rgba(0,0,0,.1);display:none;margin-top:3px;max-height:250px;max-width:400px;overflow-y:scroll;position:relative;z-index:3}.autosuggester-box-list{list-style-type:none;margin:0;padding:0}.autosuggester-box-list-item{border-bottom:1px solid var(--content-border,#eaeaea);line-height:1.2;padding:10px 15px}.autosuggester-box-list-item-value{font:300 .75rem/1.25 SFMono-Regular,Menlo,Monaco,Consolas,Liberation Mono,Courier New,monospace;margin-bottom:3px}.autosuggester-box-list-item-content{color:var(--info,#8ab858);font-size:.75rem}.autosuggester-box-list-item-active{background-color:var(--hover-row,#f5f5f5);cursor:pointer}html[data-color-scheme=dark] .autosuggester-box{background-color:var(--content-bg,#151619);border-color:#3a3d40;box-shadow:0 3px 6px #3a3d40}html[data-color-scheme=dark] .autosuggester-box-list-item{border-color:#3a3d40}html[data-color-scheme=dark] .autosuggester-box-list-item-active{background-color:#272930}
+/*# sourceMappingURL=autosuggester.1896a185.css.map*/
