--- public/autosuggester.f6693aac.css
@@ -1,2 +1,0 @@
-.autosuggester-input-mirror{left:-9999px!important;overflow:auto!important;position:absolute!important;top:0!important;white-space:pre-wrap!important;word-wrap:break-word!important}.autosuggester-box-container{position:absolute;z-index:9}.autosuggester-box{background-color:var(--content-bg);border:1px solid var(--content-border);border-radius:var(--border-radius);box-shadow:0 3px 6px 0 rgba(0,0,0,.1);display:none;margin-top:3px;max-height:250px;max-width:400px;overflow-y:scroll;position:relative;z-index:3}.autosuggester-box-list{list-style-type:none;margin:0;padding:0}.autosuggester-box-list-item{border-bottom:1px solid var(--content-border);line-height:1.2;padding:10px 15px}.autosuggester-box-list-item-value{font:300 .75rem/1.25 SFMono-Regular,Menlo,Monaco,Consolas,Liberation Mono,Courier New,monospace;margin-bottom:3px}.autosuggester-box-list-item-content{color:var(--info);font-size:.75rem}.autosuggester-box-list-item-active{background-color:var(--hover-row);cursor:pointer}
-/*# sourceMappingURL=autosuggester.f6693aac.css.map*/
