--- public/autosuggester.f90e0b3c.js
@@ -1,2 +1,0 @@
-!function(){"use strict";var t={24:function(t,e,i){i.r(e)}},e={};function i(s){var n=e[s];if(void 0!==n)return n.exports;var o=e[s]={exports:{}};return t[s](o,o.exports,i),o.exports}i.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})};!function(){i(24);var t=function(){function t(t,e){var i=this;this.cssClassInputMirror="autosuggester-input-mirror",this.cssClassInputMirrorCaret="autosuggester-input-mirror-caret",this.cssClassBox="autosuggester-box",this.cssClassBoxContainer="autosuggester-box-container",this.cssClassBoxList="autosuggester-box-list",this.cssClassBoxListItem="autosuggester-box-list-item",this.cssClassBoxListItemActive="autosuggester-box-list-item-active",this.cssClassBoxListItemValue="autosuggester-box-list-item-value",this.cssClassBoxListItemContent="autosuggester-box-list-item-content",this.boxListItems=[],this.boxListItemsVisibleIndexes=[],this.input="string"==typeof t?document.getElementById(t):t,this.tokens=e.map((function(t){return{name:"##".concat(t.name,"##"),label:t.label}})),this.tinyMCEInstance=!1,this.rgxp=new RegExp("[^ |\n]+$","i"),this.boundHideBox=function(){return i.hideBox()},this.boundKeyUp=function(t){return i.onKeyUpEvent(t)},this.boundKeyDown=function(t){return i.onKeyDownEvent(t)},this.input.autocomplete="off",setTimeout((function(){if(window.tinyMCE)try{i.tinyMCEInstance=window.tinyMCE.get(i.input.id)}catch(t){}"textarea"!==i.input.tagName.toLowerCase()||i.tinyMCEInstance||i.createInputMirror(),i.createBoxElements(),i.registerEvents()}),500)}return t.prototype.createBoxElements=function(){var t,e=this,i=document.getElementById("".concat(this.input.id,"-autosuggester"));null===(t=null==i?void 0:i.parentElement)||void 0===t||t.remove(),this.boxContainer=document.createElement("div"),this.boxContainer.className=this.cssClassBoxContainer,this.box=document.createElement("div"),this.box.id="".concat(this.input.id,"-autosuggester"),this.box.className=this.cssClassBox,this.boxList=document.createElement("ul"),this.boxList.className=this.cssClassBoxList,this.tokens.forEach((function(t,i){e.boxListItems[i]=document.createElement("li"),e.boxListItems[i].className=e.cssClassBoxListItem,e.boxListItems[i].innerHTML='<div class="'.concat(e.cssClassBoxListItemValue,'">').concat(t.name,'</div><div class="').concat(e.cssClassBoxListItemContent,'">').concat(t.label,"</div>"),e.boxList.appendChild(e.boxListItems[i]),e.boxListItemsVisibleIndexes.push(i)})),this.box.appendChild(this.boxList),this.boxContainer.appendChild(this.box),document.body.appendChild(this.boxContainer),this.boxVisible=!1},t.prototype.createInputMirror=function(){var t=this;this.inputMirror=document.createElement("div"),this.inputMirror.className=this.cssClassInputMirror,this.inputMirror.innerText=this.input.value,this.inputMirrorCaret=document.createElement("span"),this.inputMirrorCaret.className=this.cssClassInputMirrorCaret,this.inputMirrorCaret.innerHTML="&nbsp;";var e=window.getComputedStyle(this.input);["box-sizing","font-family","font-size","font-style","font-variant","font-weight","height","letter-spacing","line-height","max-height","min-height","padding-bottom","padding-left","padding-right","padding-top","text-decoration","text-indent","text-transform","width","word-spacing"].forEach((function(i){return t.inputMirror.style[i]=e[i]})),this.inputMirror.appendChild(this.inputMirrorCaret),this.input.insertAdjacentElement("afterend",this.inputMirror)},t.prototype.registerEvents=function(){var t=this;this.tinyMCEInstance||(this.input.addEventListener("keyup",this.boundKeyUp),this.input.addEventListener("keydown",this.boundKeyDown)),this.tinyMCEInstance&&("function"==typeof this.tinyMCEInstance.on?(this.tinyMCEInstance.on("keyUp",this.boundKeyUp),this.tinyMCEInstance.off("keyDown"),this.tinyMCEInstance.on("keyDown",this.boundKeyDown)):(this.tinyMCEInstance.onKeyUp.add((function(e,i){return t.boundKeyUp(i)})),this.tinyMCEInstance.onKeyDown.listeners=[],this.tinyMCEInstance.onKeyDown.add((function(e,i){return t.boundKeyDown(i)})))),this.boxListItems.forEach((function(e){e.addEventListener("mouseenter",(function(e){return t.highlightItem(t.boxListItems.indexOf(e.target))})),e.addEventListener("click",(function(e){e.preventDefault(),t.selectItem()}))})),document.addEventListener("turbo:before-cache",(function(){return t.destroy()}),{once:!0})},t.prototype.destroy=function(){var t,e;this.hideBox(),null===(t=this.boxContainer)||void 0===t||t.remove(),null===(e=this.inputMirror)||void 0===e||e.remove(),this.tinyMCEInstance||(this.input.removeEventListener("keyup",this.boundKeyUp),this.input.removeEventListener("keydown",this.boundKeyDown)),delete this.input.dataset.autosuggesterId},t.prototype.showBox=function(){if(!this.boxVisible){var t,e,i,s={left:0,top:0};this.tinyMCEInstance?(s.left=this.tinyMCEInstance.selection.getRng().getClientRects()[0].left,s.top=this.tinyMCEInstance.selection.getNode().getClientRects()[0].top+this.tinyMCEInstance.selection.getNode().getClientRects()[0].height):this.inputMirror?(t=this.input.selectionStart,i=[(e=this.input.value).substring(0,t),e.substring(t,e.length)],this.inputMirror.innerHTML="",this.inputMirror.appendChild(document.createTextNode(i[0])),this.inputMirror.appendChild(this.inputMirrorCaret),this.inputMirror.appendChild(document.createTextNode(i[1])),(s=this.getPosition(this.inputMirrorCaret,this.inputMirror)).top+=this.inputMirrorCaret.getBoundingClientRect().height):s.top=this.input.getBoundingClientRect().height,this.boxListItems.forEach((function(t){return t.classList.remove("invisible")})),this.tinyMCEInstance?this.setPosition(this.boxContainer,this.getPosition(document.getElementById("".concat(this.input.id,"_ifr")))):this.setPosition(this.boxContainer,this.getPosition(this.input)),this.currentListItemIndex=null,this.box.style.display="block",this.setPosition(this.box,s),this.box.scrollTo(0,0),this.boxVisible=!0,document.body.addEventListener("click",this.boundHideBox,{once:!0})}},t.prototype.hideBox=function(){var t=this;this.box.style.display="none",this.boxVisible=!1,this.boxListItems.forEach((function(e){return e.classList.remove(t.cssClassBoxListItemActive)})),document.body.removeEventListener("click",this.boundHideBox)},t.prototype.onKeyUpEvent=function(t){var e,i,s,n;"Escape"!==t.key&&"ArrowUp"!==t.key&&"ArrowDown"!==t.key&&(this.tinyMCEInstance?(e=(s=this.tinyMCEInstance.selection.getRng()).startContainer.wholeText,i=s.startOffset,e&&i===s.startContainer.length&&s.startContainer.textContent.length!==e.length&&(i+=e.length-s.startContainer.textContent.length)):(e=this.input.value,i=this.input.selectionStart),e?(this.filterText="",e=e.slice(0,i),(n=this.rgxp.exec(e))&&(this.showBox(),this.filterText=n[0],this.filterItems()||this.hideBox())):this.hideBox())},t.prototype.onKeyDownEvent=function(t){if(this.boxVisible){var e=this.boxListItemsVisibleIndexes.indexOf(this.currentListItemIndex);switch(t.key){case"Escape":this.hideBox();break;case"Enter":if(null===this.currentListItemIndex)return;t.preventDefault(),this.selectItem();break;case"ArrowUp":t.preventDefault(),-1===e?this.highlightItem(this.boxListItemsVisibleIndexes.length-1):e>0&&this.highlightItem(this.boxListItemsVisibleIndexes[e-1]);break;case"ArrowDown":t.preventDefault(),-1===e?this.highlightItem(this.boxListItemsVisibleIndexes[0]):e<this.boxListItemsVisibleIndexes.length-1&&this.highlightItem(this.boxListItemsVisibleIndexes[e+1])}}},t.prototype.highlightItem=function(t){var e=parseInt(window.getComputedStyle(this.box).maxHeight),i=this.box.scrollTop,s=e+i,n=this.getPosition(this.boxListItems[t],this.box).top+i,o=n+this.boxListItems[t].getBoundingClientRect().height;null!==this.currentListItemIndex&&this.boxListItems[this.currentListItemIndex].classList.remove(this.cssClassBoxListItemActive),this.boxListItems[t].classList.add(this.cssClassBoxListItemActive),this.currentListItemIndex=t,o>=s?this.box.scrollTo(0,o-e>0?o-e:0):n<i&&this.box.scrollTo(0,n)},t.prototype.selectItem=function(){var t,e,i,s=this.tokens[this.currentListItemIndex].name;this.filterText.length>0&&(s=s.substring(this.filterText.length,s.length)),this.tinyMCEInstance?this.tinyMCEInstance.selection.setContent(s):(t=this.input.value,i=(e=this.input.selectionStart)+s.length,this.input.value=t.slice(0,e)+s+t.slice(e,t.length),this.input.setSelectionRange(i,i)),this.hideBox()},t.prototype.filterItems=function(){var t=this,e=new RegExp("^".concat(this.filterText,"(?!$)"),"i");return this.boxListItems.forEach((function(i,s){e.test(t.tokens[s].name)?(t.boxListItems[s].classList.remove("invisible"),t.boxListItemsVisibleIndexes.includes(s)||t.boxListItemsVisibleIndexes.push(s)):(t.boxListItems[s].classList.add("invisible"),t.boxListItemsVisibleIndexes.includes(s)&&t.boxListItemsVisibleIndexes.splice(t.boxListItemsVisibleIndexes.indexOf(s),1))})),this.boxListItemsVisibleIndexes.length>0},t.prototype.getPosition=function(t,e){void 0===e&&(e=null);var i=t.getBoundingClientRect(),s=i.top+window.scrollY,n=i.left+window.scrollX;if(null!==e){var o=this.getPosition(e);s-=o.top,n-=o.left}return{top:s,left:n}},t.prototype.setPosition=function(t,e){t.style.left="".concat(e.left,"px"),t.style.top="".concat(e.top,"px")},t}();function e(){document.querySelectorAll("script[data-notification-center-auto-suggester]").forEach((function(e){var i=JSON.parse(e.textContent),s=i.input,n=i.tokens,o=document.getElementById(s);o&&!o.dataset.autosuggesterId&&(new t(o,n),o.dataset.autosuggesterId="initialized")}))}document.addEventListener("DOMContentLoaded",e),document.addEventListener("turbo:load",e),document.addEventListener("turbo:render",e)}()}();
-//# sourceMappingURL=autosuggester.f90e0b3c.js.map
