+++ public/entrypoints.json
@@ -2,18 +2,10 @@
   "entrypoints": {
     "autosuggester": {
       "css": [
-        "/bundles/terminal42notificationcenter/autosuggester.f6693aac.css"
+        "/bundles/terminal42notificationcenter/autosuggester.1896a185.css"
       ],
       "js": [
-        "/bundles/terminal42notificationcenter/autosuggester.f90e0b3c.js"
-      ]
-    },
-    "legacy/autosuggester": {
-      "css": [
-        "/bundles/terminal42notificationcenter/legacy/autosuggester.f5be80e8.css"
-      ],
-      "js": [
-        "/bundles/terminal42notificationcenter/legacy/autosuggester.8b356cab.js"
+        "/bundles/terminal42notificationcenter/autosuggester.4dabb3ea.js"
       ]
     },
     "backend": {
