--- public/legacy/autosuggester.f5be80e8.css
@@ -1,2 +1,0 @@
-.autosuggester-input-mirror{left:-9999px!important;overflow:auto!important;position:absolute!important;top:0!important;white-space:pre-wrap!important;word-wrap:break-word!important}.autosuggester-box-container{position:absolute;z-index:9}.autosuggester-box{background-color:#fff;border:1px solid #bbb;border-radius:0 0 3px 3px;box-shadow:0 3px 6px #ccc;display:none;margin-top:4px;max-height:150px;overflow-y:scroll;position:relative;width:318px;z-index:3}.autosuggester-box-list{list-style-type:none;margin:0;padding:0}.autosuggester-box-list-item{border-bottom:1px solid #eaeaea;padding:8px}.autosuggester-box-list-item-value{font-weight:700}.autosuggester-box-list-item-active{background-color:#f5f5f5;cursor:pointer}.autosuggester-box-list-item-active .autosuggester-box-list-item-value{color:#8ab858}html[data-color-scheme=dark] .autosuggester-box{background-color:#151619;border-color:#3a3d40;box-shadow:0 3px 6px #3a3d40}html[data-color-scheme=dark] .autosuggester-box-list-item{border-color:#3a3d40}html[data-color-scheme=dark] .autosuggester-box-list-item-active{background-color:#272930}
-/*# sourceMappingURL=autosuggester.f5be80e8.css.map*/
