+++ public/manifest.json
@@ -1,14 +1,10 @@
 {
-  "autosuggester.css": "/bundles/terminal42notificationcenter/autosuggester.f6693aac.css",
-  "autosuggester.js": "/bundles/terminal42notificationcenter/autosuggester.f90e0b3c.js",
-  "legacy/autosuggester.css": "/bundles/terminal42notificationcenter/legacy/autosuggester.f5be80e8.css",
-  "legacy/autosuggester.js": "/bundles/terminal42notificationcenter/legacy/autosuggester.8b356cab.js",
+  "autosuggester.css": "/bundles/terminal42notificationcenter/autosuggester.1896a185.css",
+  "autosuggester.js": "/bundles/terminal42notificationcenter/autosuggester.4dabb3ea.js",
   "backend.css": "/bundles/terminal42notificationcenter/backend.023485eb.css",
   "images/award.svg": "/bundles/terminal42notificationcenter/images/award.8c4060af.svg",
   "images/bell.svg": "/bundles/terminal42notificationcenter/images/bell.49b754b0.svg",
-  "autosuggester.f6693aac.css.map": "/bundles/terminal42notificationcenter/autosuggester.f6693aac.css.map",
-  "autosuggester.f90e0b3c.js.map": "/bundles/terminal42notificationcenter/autosuggester.f90e0b3c.js.map",
-  "autosuggester.f5be80e8.css.map": "/bundles/terminal42notificationcenter/legacy/autosuggester.f5be80e8.css.map",
-  "autosuggester.8b356cab.js.map": "/bundles/terminal42notificationcenter/legacy/autosuggester.8b356cab.js.map",
+  "autosuggester.1896a185.css.map": "/bundles/terminal42notificationcenter/autosuggester.1896a185.css.map",
+  "autosuggester.4dabb3ea.js.map": "/bundles/terminal42notificationcenter/autosuggester.4dabb3ea.js.map",
   "backend.023485eb.css.map": "/bundles/terminal42notificationcenter/backend.023485eb.css.map"
 }
