+++ webpack.config.js
@@ -1,11 +1,8 @@
 const Encore = require('@terminal42/contao-build-tools');
 
-module.exports = Encore()
+module.exports = Encore('assets')
     .setOutputPath('public/')
     .setPublicPath('/bundles/terminal42notificationcenter')
-    .addEntry('autosuggester', './assets/autosuggester.ts')
-    .addEntry('legacy/autosuggester', './assets/legacy/autosuggester.ts')
     .addStyleEntry('backend', './assets/backend.scss')
-    .enableTypeScriptLoader()
     .getWebpackConfig()
 ;
